+++ EvaluacionAcademia.NET/DTOs/AccountFiduciaryDto.cs
@@ -6,7 +6,7 @@
 		public string CBU { get; set; }
 		public string Alias { get; set; }
 		public string AccountNumber { get; set; }
-		public float BalancePeso { get; set; }
-		public float BalanceUsd { get; set; }
+		public decimal BalancePeso { get; set; }
+		public decimal BalanceUsd { get; set; }
 	}
 }
