+++ EvaluacionAcademia.NET/DataAccess/Repositories/Interfaces/IAccountFiduciaryRepository.cs
@@ -4,7 +4,7 @@ namespace EvaluacionAcademia.NET.DataAccess.Repositories.Interfaces
 {
 	public interface IAccountFiduciaryRepository : IRepository<AccountFiduciary>
 	{
-		Task<float> GetBalancePeso(int accountId);
-		Task<float> GetBalanceUsd(int accountId);
+		Task<decimal> GetBalancePeso(int accountId);
+		Task<decimal> GetBalanceUsd(int accountId);
 	}
 }
